Move remaining linking logic from UI to AppThunks
